Crate rust_higher_kinded_types

This is a simple library designed to showcase a prospective design for a family of zipper-like graph traversal traits. These traits may be implemented over varying graph implementations (adjacency list, adjacency matrix, lazily constructed infinite graph, actual zippers, etc.).

Basic use of them goes something like this, assuming some type V that implements nav_types::Vertex:

let vertex: V = ...;
// List children of vertex. vertex.out_edges() returns a type that
// implements nav_types::OutEdgeCollection.
for edge in vertex.out_edges().iter() {
  // edge implements nav_types::Edge.
  println!("vertex with data {:?} has child edge with data {:?}",
           vertex.data(), edge.data())
}

These traits are backed by a borrow of an arbitrary graph structure, and Rust's ability to track lifetimes means that that may be done safely.

Modules

mut_types
Types that can traverse the graph and may modify its structure. These types are backed by a mutable borrow of graph data, so multiple active (unborrowed) instances cannot exist.

nav_types
Types that can traverse the graph and cannot modify its structure. These types are backed by a read-only borrow of graph data, so multiple instances of them may coexist.

Traits

BoundedIterator
When iterating over components of a graph structure, we need to give a name to the borrow of the underlying graph. This trait does so.

GraphTypes
Base types that the graph defines a structure over.